[REFACT] 작가 회원가입 서비스 분리
