+++ src/main/java/net/catsnap/domain/photographer/service/PhotographerService.java
@@ -1,7 +1,6 @@
 package net.catsnap.domain.photographer.service;
 
 import lombok.RequiredArgsConstructor;
-import net.catsnap.domain.photographer.converter.PhotographerConverter;
 import net.catsnap.domain.photographer.document.PhotographerReservationLocation;
 import net.catsnap.domain.photographer.document.PhotographerReservationNotice;
 import net.catsnap.domain.photographer.document.PhotographerSetting;
@@ -12,11 +11,8 @@ import net.catsnap.domain.photographer.repository.PhotographerRepository;
 import net.catsnap.domain.photographer.repository.PhotographerReservationLocationRepository;
 import net.catsnap.domain.photographer.repository.PhotographerReservationNoticeRepository;
 import net.catsnap.domain.photographer.repository.PhotographerSettingRepository;
-import net.catsnap.domain.reservation.service.PhotographerReservationService;
 import net.catsnap.global.Exception.authority.ResourceNotFoundException;
-import net.catsnap.global.Exception.photographer.DuplicatedPhotographerException;
 import net.catsnap.global.security.contextholder.GetAuthenticationInfo;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,51 +21,10 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class PhotographerService {
 
-    private final PasswordEncoder passwordEncoder;
     private final PhotographerRepository photographerRepository;
     private final PhotographerSettingRepository photographerSettingRepository;
-    private final PhotographerConverter photographerConverter;
     private final PhotographerReservationNoticeRepository photographerReservationNoticeRepository;
     private final PhotographerReservationLocationRepository photographerReservationLocationRepository;
-
-    private final PhotographerReservationService photographerReservationService;
-
-    public void singUp(PhotographerRequest.PhotographerSignUp photographerSignUp) {
-        photographerRepository.findByIdentifier(photographerSignUp.getIdentifier())
-            .ifPresent(Photographer -> {
-                throw new DuplicatedPhotographerException("이미 존재하는 아이디입니다.");
-            });
-
-        String encodedPassword = passwordEncoder.encode(photographerSignUp.getPassword());
-        Photographer photographer = photographerConverter.photographerSignUpToPhotographer(
-            photographerSignUp, encodedPassword);
-        photographerRepository.save(photographer);
-
-
-        /*
-         * 작가 회원가입 시 초기화 작업
-         */
-        // weekdayReservationTimeMapping 생성 (예약을 형식을 요일에 매핑하는 테이블 생성)
-        photographerReservationService.createJoinedPhotographerReservationTimeFormat(photographer);
-        // photographerSetting 초기화
-        initializeSineUpPhotographer(photographer.getId());
-        // todo : 이용약관 동의 여부 확인
-    }
-
-    /*
-     * 작가 회원가입 시 초기화 작업
-     */
-    public void initializeSineUpPhotographer(Long photographerId) {
-        PhotographerSetting photographerSetting = PhotographerSetting.builder()
-            .photographerId(photographerId)
-            .autoReservationAccept(false)
-            .enableOverBooking(false)
-            .preReservationDays(14L)
-            .build();
-        photographerSettingRepository.save(photographerSetting);
-        photographerReservationNoticeRepository.save("", photographerId);
-        photographerReservationLocationRepository.save("", photographerId);
-    }
 
     public PhotographerSetting getPhotographerSetting() {
         Long photographerId = GetAuthenticationInfo.getUserId();
